MSVC: Suppress link warning
Recent versions of MSVC have starting issuing warnings when the `.dbg` file
associated with a library isn't found when linking Debug code.  I thought this
issue was restricted to libeospac.lib, but it turns out other TPLs also
cause this warning to be emitted.

I'm pulling the change to `cdi_eospac/CMakeLists.txt` that I made in #649 and
replacing it by moving the `ignore` link flag to the global `LINK_OPTIONS`
CMake variable.