Cleanup in quadrature.
Background:

A couple of weeks ago I was debugging a test failure that ran through
quadrature.  In the end, there was nothing wrong with quadrature.  However,
since I had touched a bunch of files in this package, a PR seemed to be in
order.  Changes are mostly code cleanup.

This debugging session also required some build system changes to allow the
MSVC debugger to find all necessary symbols.  This PR includes these changes
that are related to Redmine ticket 1765.

Changes:

+ For ds++ and quadrature, install the MSVC .pdb files if they are available.
+ Remove namespace injection of `std::` by removing `using namespace std;` from
 `.hh` files.  Similarly, remove namespece injection for
 `rtt_mesh_element::Geometry`.
+ Spelling fixes, comment reindentation, removal of unnecessary `DLL_PUBLIC_`
  decoration and removal of left over debug print statements.
+ Remove empty file `Ordinate.cc`.
+ Eliminate a build warning by adding a copy constructor to `Ordinate_Set`.
+ Removal of double underscore in CPP names (in a few places).
+ Replace `NULL` with `nullptr`.